fix: data type

+++ training_part/data_loader.py
@@ -103,8 +103,8 @@ class SPRDataset(Dataset):
             label_mask = augmented["mask"]
             
         if idx % 10 == 0:
-            Image.fromarray(img.numpy().transpose(1, 2, 0)).save(f"temp_images/img_{str(idx).zfill(5)}.png")
-            Image.fromarray(label_mask.numpy().transpose(1, 2, 0)).save(f"temp_images/lab_{str(idx).zfill(5)}.png")
+            Image.fromarray(img.numpy().transpose(1, 2, 0).astype(np.uint8)).save(f"temp_images/img_{str(idx).zfill(5)}.png")
+            Image.fromarray(label_mask.numpy().transpose(1, 2, 0).astype(np.uint8)).save(f"temp_images/lab_{str(idx).zfill(5)}.png")
         return img, label_mask
     
     def _read_paths(self):
